requirements.txt
# requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2                 # Async MongoDB driver
pymongo==4.6.0
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Telegram Bot
python-telegram-bot==20.7

# CLI Interface
click==8.1.7
colorama==0.4.6
rich==13.7.0                 # Rich text and beautiful formatting

# Security & Authentication
cryptography==45.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2               # For testing HTTP clients

# Development Tools
black==23.11.0              # Code formatting
flake8==6.1.0               # Linting
mypy==1.7.1                 # Type checking

# Async Support
asyncio-mqtt==0.16.1        # For async MQTT if needed

# Math/Statistics (for backtesting)
scipy==1.11.4
scikit-learn==1.3.2

# Web UI Templates
jinja2==3.1.2
python-socketio==5.10.0     # For real-time web updates

# Email Support (for notifications)
aiosmtplib==3.0.1

# Additional Utilities
schedule==1.2.0             # For scheduled tasks
psutil==5.9.6               # System monitoring
requests==2.31.0            # HTTP requests (backup to httpx)

# Optional: ML/AI Features (for future signal analysis)
# tensorflow==2.15.0
# scikit-learn==1.3.2